+++ src/tado/tado-config.js
@@ -582,9 +582,8 @@ export default {
                             ? 'zone-faucet'
                             : 'zone-switch';
 
-                    config.subtype = zone.boilerTempSupport && zone.accTypeBoiler !== 'FAUCET'
-                      ? 'zone-heatercooler-boiler'
-                      : config.subtype;
+                    // boilerTempSupport always takes precedence over accTypeBoiler
+                    config.subtype = zone.boilerTempSupport ? 'zone-heatercooler-boiler' : config.subtype;
 
                     config.zoneId = zone.id;
                     config.type = zone.type;